requirements.txt
bibtexparser==0.6.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-autoslug==1.9.3
django-better-filter-widget==0.4
django-citations==0.3
django-countries==3.4.1
django-extensions==1.7.4
django-grappelli==2.7.2
django-markdown==0.8.4
django-select-multiple-field==0.4.2
django-smart-selects==1.2.0
django-toolbelt==0.0.1
gunicorn==19.6
djangorestframework==3.5.2
Markdown==2.6.5
psycopg2==2.6.1
pytz==2015.7
six==1.10.0
static3==0.6.1
Unidecode==0.4.18
whitenoise==3.2.2
